pyproject.toml
[tool.poetry]
name = "drf_api"
version = "0.1.0"
description = ""
authors = ["HamzaQahoush <hamza.qah@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
